requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.10
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==1.26.4
nvidia-nccl-cu12==2.28.3
orjson==3.11.3
pandas==2.2.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.1.0
rich-toolkit==0.15.1
scikit-learn==1.5.1
scipy==1.16.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
threadpoolctl==3.6.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xgboost==2.1.1
